+++ buildSrc/src/main/kotlin/common/BuildProfile.kt
@@ -127,7 +127,7 @@ object BuildProfiles {
             platformVersionCode = "231",
             pluginSinceBuild = "231",
             pluginUntilBuild = "231.*",
-            versionToRunPluginVerifier = "2023.1.4",
+            versionToRunPluginVerifier = "2023.1.5",
             kotlinTarget = KotlinVersion.KOTLIN_1_8.version,
             javaVersion = JavaVersion.VERSION_17.majorVersion
         ),
@@ -135,7 +135,7 @@ object BuildProfiles {
         Profiles.p232 to BuildProfile(
             profile = Profiles.p232,
             // platformVersion is Intellij IDEA Community and Ultimate
-            platformVersion = "2023.2.1",
+            platformVersion = "2023.2.2",
             riderVersion = "2023.2.1",
             pycharmVersion = "2023.2.1",
             riderResharperVersion = "2023.2.0",
@@ -144,7 +144,7 @@ object BuildProfiles {
             platformVersionCode = "232",
             pluginSinceBuild = "232",
             pluginUntilBuild = "232.*",
-            versionToRunPluginVerifier = "2023.2.1",
+            versionToRunPluginVerifier = "2023.2.2",
             kotlinTarget = KotlinVersion.KOTLIN_1_8.version,
             javaVersion = JavaVersion.VERSION_17.majorVersion
         ),