+++ ide-common/src/main/kotlin/org/digma/intellij/plugin/docker/DigmaInstallationStatus.kt
@@ -1,0 +1,16 @@
+package org.digma.intellij.plugin.docker
+
+
+enum class DigmaInstallationType {
+    localEngine, dockerCompose, dockerDesktop
+}
+
+
+data class DigmaInstallationStatus(
+    val connection: ConnectionStatus,
+    val runningDigmaInstances: List<DigmaInstallationType>,
+)
+
+enum class ConnectionType { local, remote }
+
+data class ConnectionStatus(val type: ConnectionType?, val status: Boolean)