+++ src/main/java/org/digma/intellij/plugin/emvironment/EnvironmentChangeHandler.java
@@ -8,6 +8,7 @@ import org.digma.intellij.plugin.errorreporting.ErrorReporter;
 import org.digma.intellij.plugin.log.Log;
 import org.digma.intellij.plugin.psi.LanguageService;
 import org.digma.intellij.plugin.psi.SupportedLanguages;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 
@@ -31,7 +32,7 @@ public class EnvironmentChangeHandler implements EnvironmentChanged {
     //environmentChanged must run in a background thread.
     //when fired by the Environment object it is on background
     @Override
-    public void environmentChanged(String newEnv, boolean refreshInsightsView) {
+    public void environmentChanged(@Nullable String newEnv, boolean refreshInsightsView) {
 
         try {